Refactor checkIfQuestionMeetsCriteria() to reuse
