+++ packages/desktop/app/utils/survey.js
@@ -1,7 +1,7 @@
 // Much of this file is duplicated in `packages/mobile/App/ui/components/Forms/SurveyForm/helpers.ts`
 import React from 'react';
 import * as yup from 'yup';
-import { checkJSONCriteria } from '@tamanu/shared/utils/criteria';
+import { checkJSONVisibilityCriteria } from '@tamanu/shared/utils';
 import { intervalToDuration, parseISO } from 'date-fns';
 
 import {
@@ -92,14 +92,7 @@ export function checkVisibility(component, values, allComponents) {
   if (!visibilityCriteria) return true;
 
   try {
-    const valuesByCode = Object.entries(values).reduce((acc, [key, val]) => {
-      const matchingComponent = allComponents.find(x => x.dataElement.id === key);
-      if (matchingComponent) {
-        acc[matchingComponent.dataElement.code] = val;
-      }
-      return acc;
-    }, {});
-    return checkJSONCriteria(visibilityCriteria, allComponents, valuesByCode);
+    return checkJSONVisibilityCriteria(component, allComponents, values);
   } catch (error) {
     // eslint-disable-next-line no-console
     console.warn(`Error parsing visilbity criteria as JSON, using fallback.
