+++ packages/shared/src/utils/fields.js
@@ -40,6 +40,56 @@ function compareData(dataType, expected, given) {
   return false;
 }
 
+export function checkJSONVisibilityCriteria(component, allComponents, values) {
+  const { visibilityCriteria } = component;
+  // nothing set - show by default
+  if (!visibilityCriteria) return true;
+
+  const criteriaObject = JSON.parse(visibilityCriteria);
+
+  if (!criteriaObject) {
+    return true;
+  }
+
+  const { _conjunction: conjunction, hidden: _, ...restOfCriteria } = criteriaObject;
+  if (Object.keys(restOfCriteria).length === 0) {
+    return true;
+  }
+
+  const checkIfQuestionMeetsCriteria = ([questionCode, answersEnablingFollowUp]) => {
+    const matchingComponent = allComponents.find(x => x.dataElement?.code === questionCode);
+    const value = values[questionCode] || values[matchingComponent.dataElement.id]; // Usages: central, desktop, shared. Only desktop use dataElementId to fetch value.
+    if (answersEnablingFollowUp.type === 'range') {
+      if (!value && value !== 0) return false;
+      const { start, end } = answersEnablingFollowUp;
+
+      if (!start) return value < end;
+      if (!end) return value >= start;
+      if (inRange(parseFloat(value), parseFloat(start), parseFloat(end))) {
+        return true;
+      }
+      return false;
+    }
+
+    const isMultiSelect =
+      matchingComponent?.dataElement?.type === PROGRAM_DATA_ELEMENT_TYPES.MULTI_SELECT;
+
+    if (Array.isArray(answersEnablingFollowUp)) {
+      return isMultiSelect
+        ? (value?.split(', ') || []).some(selected => answersEnablingFollowUp.includes(selected))
+        : answersEnablingFollowUp.includes(value);
+    }
+
+    return isMultiSelect
+      ? value?.includes(answersEnablingFollowUp)
+      : answersEnablingFollowUp === value;
+  };
+
+  return conjunction === 'and'
+    ? Object.entries(restOfCriteria).every(checkIfQuestionMeetsCriteria)
+    : Object.entries(restOfCriteria).some(checkIfQuestionMeetsCriteria);
+}
+
 /**
  * IMPORTANT: We have 4 other versions of this method:
  *
@@ -55,7 +105,7 @@ function checkVisibilityCriteria(component, allComponents, values) {
   const { visibilityCriteria } = component;
 
   try {
-    return checkJSONCriteria(visibilityCriteria, allComponents, values);
+    return checkJSONVisibilityCriteria(component, allComponents, values);
   } catch (error) {
     log.warn(
       `Error parsing JSON visibility criteria for ${component.dataElement?.code}, using fallback.\nError message: ${error.message}`,
