+++ packages/shared/src/utils/index.js
@@ -20,3 +20,4 @@ export * from './getCovidClearanceCertificateFilter';
 export * from './getLabTestsFromLabRequests';
 export { renameObjectKeys } from './renameObjectKeys';
 export * from './objectToCamelCase';
+export { checkJSONVisibilityCriteria } from './fields';
